Improvements for Ruler and Cytoband
Improved sizing of these components; may end up re-doing that entirely
to handle resize but this lays them out much more like the Swing UI.
Added a border to the Ruler; this needs to be drawn on the canvas
rather than done via CSS, otherwise it doesn't work.
Fixed a glaring swap of Height for Width in constructing the Scene.
Added lots of size-debugging log statements; will remove these later.
Switched some component sizing from int to double.

+++ src/org/broad/igv/prefs/IGVPreferences.java
@@ -528,10 +528,10 @@ public class IGVPreferences {
 
         if (applicationBounds != null) {
             String[] values = applicationBounds.split(",");
-            int x = Integer.parseInt(values[0]);
-            int y = Integer.parseInt(values[1]);
-            int width = Integer.parseInt(values[2]);
-            int height = Integer.parseInt(values[3]);
+            double x = Double.parseDouble(values[0]);
+            double y = Double.parseDouble(values[1]);
+            double width = Double.parseDouble(values[2]);
+            double height = Double.parseDouble(values[3]);
 
             if (width == 0 || height == 0) {
                 return null;  // Don't know bounds
